Added new helper method to add attribute to existing logger. Better grouped existing code
