+++ util/log.go
@@ -15,6 +15,21 @@ func init() {
 	slog.SetDefault(slog.New(slog.NewJSONHandler(os.Stderr, slogOpts)))
 }
 
+type contextKey string
+
+const (
+	loggerKey contextKey = "logger"
+)
+
+func LoggerFromContext(ctx context.Context) *slog.Logger {
+	if l := ctx.Value(loggerKey); l == nil {
+		slog.Warn("Using default slog as context does not have logger info")
+		return slog.Default()
+	} else {
+		return l.(*slog.Logger)
+	}
+}
+
 func NewRequestSetup(ctx context.Context, operation string, customAttributes ...slog.Attr) (*slog.Logger, context.Context) {
 	defaults := []any{slog.String("requestID", uuid.New().String()), slog.String("operation", operation)}
 
@@ -23,5 +38,17 @@ func NewRequestSetup(ctx context.Context, operation string, customAttributes ...
 	}
 
 	l := slog.With(defaults...)
+	return l, context.WithValue(ctx, loggerKey, l)
+}
+
+func AddAttribute(ctx context.Context, customAttributes ...slog.Attr) (*slog.Logger, context.Context) {
+	newAttributes := []any{}
+
+	for _, customAttribute := range customAttributes {
+		newAttributes = append(newAttributes, customAttribute)
+	}
+
+	l := LoggerFromContext(ctx)
+	l = l.With(newAttributes...)
 	return l, context.WithValue(ctx, loggerKey, l)
 }
